add lecture form v3
