--- lib/DashboardItem.dart
@@ -1,34 +1,0 @@
-import 'package:flutter/material.dart';
-
-class DashboardItem extends StatelessWidget {
-  final String title;
-  final int count;
-
-  DashboardItem({required this.title, required this.count});
-
-  @override
-  Widget build(BuildContext context) {
-    return Card(
-      elevation: 4,
-      margin: EdgeInsets.all(8),
-      child: Padding(
-        padding: const EdgeInsets.all(16.0),
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            Text(
-              title,
-              style: TextStyle(fontSize: 18, fontWeight: FontWeight.bold),
-              textAlign: TextAlign.center,
-            ),
-            SizedBox(height: 10),
-            Text(
-              count.toString(),
-              style: TextStyle(fontSize: 24, fontWeight: FontWeight.bold, color: Colors.blue),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
